v0.9.21 — Fix sync overwriting user's gem values on reload
ROOT CAUSE: initialSync() pulled task templates from Supabase on every
page load, overwriting local gem_value edits with stale Supabase data
(which had int column truncating decimals to 1).

FIX:
- First load (no local data): full pull from Supabase (unchanged)
- Subsequent loads: only sync completions, NEVER overwrite task
  templates or store items. Local edits are authoritative.
- Also processes queued writes on reconnect
- Schema: gem_value changed from int to numeric for fractional support

NOTE: User needs to run in Supabase SQL Editor:
  ALTER TABLE task_templates ALTER COLUMN gem_value TYPE numeric;

+++ supabase-schema.sql
@@ -24,7 +24,7 @@ create table task_templates (
   title text not null,
   task_type text not null check (task_type in ('daily', 'weekly')),
   parent_id uuid references task_templates(id) on delete cascade,
-  gem_value int default 1,
+  gem_value numeric default 1,
   bonus_gems int default 0,
   sort_order int default 0,
   active boolean default true,
